feat: SURVEY_QUESTION_TYPE 이동 (#95)
