+++ src/models/survey/surveyModels.ts
@@ -1,3 +1,17 @@
+export const SURVEY_QUESTION_TYPE = {
+  Radio: 'Radio',
+  Check: 'Check',
+  Text: 'Text',
+} as const;
+
+type SurveyQuestionsType = (typeof SURVEY_QUESTION_TYPE)[keyof typeof SURVEY_QUESTION_TYPE];
+
+export interface SurveyQuestion {
+  id: number;
+  type: SurveyQuestionsType;
+  text: string;
+}
+
 export interface CommonProps {
   currentStep: number;
   totalSteps: number;
